1.修复主串口通信句柄为0的bug 2.电子秤任务修复parse_pdu数据长度计算b错误崩溃的bug 3.修复上位机输出01 11不带校验，因为版本为1冲突的bug 4.所有任务栈改为256*4字节

+++ iw_controller/board/user/tasks/scale_task.c
@@ -300,7 +300,7 @@ static int receive_adu(const int handle,uint8_t *adu,uint32_t wait_timeout)
 * @return 
 * @note
 */
-static int parse_pdu(const uint8_t *pdu,uint8_t size,const uint8_t addr,const uint8_t code,uint8_t *value)
+static int parse_pdu(const uint8_t *pdu,int size,const uint8_t addr,const uint8_t code,uint8_t *value)
 {
     int rc;
     uint8_t opt_code;
